code refactor oraz monetki
