+++ screen/map_module.py
@@ -1,62 +1,14 @@
-import settings
-from character import *
-import math
-
 import pygame
-
-class GameScreen():
-
-
-    """
-    Klasa po któej dziedziczyć będą pozostałe ekrany, tj Mapa, Garaż czy Menu/Options
-
-    w,h - rozdzielczość ekranu
-    screen - odpalenie pokazywania ekranu w pygame
-
-    self.font = pygame.font.Font("textures/font.ttf", 24)
-    """
-    
-    
-    def __init__(self, w, h):
-        self.w = w
-        self.h = h
-        self.screen = pygame.display.set_mode((self.w, self.h))
-        self.scroll = 0
-        self.background = pygame.image.load('textures/tlogry.png').convert_alpha()
-        # ile obrazkow ma byc w bufforze
-        self.tiles = math.ceil(self.h / self.background.get_height()) + 2
-        self.font = pygame.font.Font("textures/font.ttf", 24)
-        
-
-    
-
-    def rescale_background(self):
-        self.background = pygame.transform.scale(self.background, (self.w, self.h)) 
-        
-        
+import math
+import random
 
 
+import settings
 
-class GameOverScreen(GameScreen):
-    def __init__(self, w, h):
-        super().__init__(w, h)
-        
-        self.text1 = self.font.render("press SPACE to start", True, 'black')
-        self.text_rect1 = self.text1.get_rect(center = (self.w//2, self.h//2))
-        self.text2 = self.font.render("press m to go back to menu", True, 'black')
-        self.text_rect2 = self.text2.get_rect(center = (self.w//2, self.h//2 + 200))
-
-    def display_bg(self):
-        self.screen.fill('white')
-        self.screen.blit(self.text1, self.text_rect1)
-        self.screen.blit(self.text2, self.text_rect2)
-        
-        
-    def display_score(self, score):
-        score_txt = self.font.render(f'score: {score}', True, 'black')
-        score_txt_rect = score_txt.get_rect(center = (self.w//2, 200))
-        self.screen.blit(score_txt, score_txt_rect)
-
+from screen.game_screen_class import GameScreen
+from character.player import Player
+from character.obstacle import StaticObstacle, DynamicObstacle
+from character.coin import Coin
 
 
 
@@ -97,19 +49,23 @@ class Map(GameScreen):
         self.hp1 = pygame.transform.rotozoom(self.hp1, 0, 0.15)
         self.hp2 = pygame.transform.rotozoom(self.hp2, 0, 0.15)
         self.hp3 = pygame.transform.rotozoom(self.hp3, 0, 0.15)
+
+        self.scroll = 0
+
+        self.coins = []
         
         
     def display_bg(self):
         # wyświetl tło 
         #self.background = pygame.transform.scale(self.background, (self.w, self.h))
 
-        self.i = 0
-        while (self.i < self.tiles):
+        i = 0
+        while (i < self.tiles):
             self.screen.blit(self.background, (0
-                             , self.background.get_height() * -self.i + self.scroll))
-            self.i += 1
+                             , self.background.get_height() * - i + self.scroll))
+            i += 1
 
-        self.scroll += 5
+        self.scroll += 5 * self.game_speed
 
         # resetowanie scrolla
         if abs(self.scroll) > self.background.get_height():
@@ -126,17 +82,23 @@ class Map(GameScreen):
 
 
 
-    def display_score(self):
+
+
+    def display_score_and_money(self):
         self.calculate_score()
         self.score_txt = self.font.render(f'score: {self.score}', True, 'gold')
         self.score_txt_rect = self.score_txt.get_rect(topleft = (10, 10))
+        self.money_txt = self.font.render(f'coins: {self.player1.game_money}', True, 'gold')
+        self.money_txt_rect = self.money_txt.get_rect(topleft = (10, 30))
+
         self.screen.blit(self.score_txt, self.score_txt_rect)
+        self.screen.blit(self.money_txt, self.money_txt_rect)
         
     def increase_speed(self):
         # przyspiesz grę co 3 sekundy -- wartości można zmienić
         if pygame.time.get_ticks() % 300 == 0:
-            self.game_speed *= 1.05
-            self.player1.dx *= 1.05
+            self.game_speed *= 1.01
+            self.player1.dx *= 1.01
 
 
     def update_characters(self):
@@ -145,10 +107,14 @@ class Map(GameScreen):
         # jeżeli wystąpi kolizja to game_over = True
         self.increase_speed()
         self.add_obstacle()
+        self.add_coin()
         self.game_over = self.check_for_obs_collision()
         
+        
+        self.collect_coin()
         self.update_player()
         self.update_obstacles()
+        self.update_coins()
                 
     def update_player(self):
         # print(self.check_for_border_collision())
@@ -180,23 +146,56 @@ class Map(GameScreen):
                     self.obstacles.remove(obstacle)
                     del obstacle
 
+    def update_coins(self):
+        for coin in self.coins:
+            if coin:    
+                coin.move(self.game_speed)
+                coin.rect.y = coin.y
+                if coin.y > self.h + 100:
+                    # gdy przeszkoda wyjdzie poza ekran jest usuwana z listy i obiekt też jest usuwany
+                    self.coins.remove(coin)
+                    del coin
+
+    def check_for_obs_coin_spawn_collision(self, to_be_added) -> bool:
+        # sprawdza czy moneta lub przeszkoda do dodania
+        # po dodaniu kolidowałaby z już istniejącą 
+        for obstacle in self.obstacles:
+            if to_be_added.rect.y <= obstacle.rect.y + 130 and to_be_added.rect.x == obstacle.rect.x:
+                return True
+        for coin in self.coins:
+            if to_be_added.rect.y <= coin.rect.y - 45 and to_be_added.rect.x == coin.rect.x:
+                return True
+        return False
+
     def add_obstacle(self):
         # dodaje przeszkodę we w miarę losowym momencie, nie może być póki co więcej niż 5 na ekranie
         # sprawdza czy w miejscu gdzie ma się pojawić przeszkoda występuje już jakaś inna przeszkoda
-        if random.randint(1, 20) % 19 == 0 and len(self.obstacles) < 5:
-            if random.randint(1,11) % 10 == 0:
+        if random.randint(1, 100) % 33 == 0 and len(self.obstacles) < 5:
+            if random.randint(1,9) % 8 == 0:
                 obs = StaticObstacle()
             else:
                 obs = DynamicObstacle()
-            collides = False
-            for obstacle in self.obstacles:
-                if obs.rect.y <= obstacle.rect.y - 130 or obs.rect.x == obstacle.rect.x:
-                    collides = True
-                    break
-            if collides:
+            if self.check_for_obs_coin_spawn_collision(obs):
                 del obs    
             else:
                 self.obstacles.append(obs)
+
+    def add_coin(self):
+        # dodanie monety
+        if random.randint(1, 100) % 97 == 0 and len(self.coins) < 4:
+            coin = Coin()
+            if self.check_for_obs_coin_spawn_collision(coin):
+                del coin    
+            else:
+                self.coins.append(coin)
+
+    def collect_coin(self):
+        # zbiera monetkę jeżeli auto się z nią zderzy
+        for coin in self.coins:
+            if pygame.Rect.colliderect(coin.rect, self.player1.rect):
+                self.player1.game_money += 1
+                self.coins.remove(coin)
+                del coin
     
 
     def check_for_obs_collision(self) -> bool:
@@ -205,6 +204,8 @@ class Map(GameScreen):
             if pygame.Rect.colliderect(obstacle.rect, self.player1.rect):
                 self.player1.hp -= 1
                 self.obstacles.remove(obstacle)
+                del obstacle
+                
                 # self.game_over = True
                 return self.checking_hp()
 
@@ -223,8 +224,12 @@ class Map(GameScreen):
         # debug:
         self.display_debug_rect()
         
+        for coin in self.coins:
+            coin.display_character(self.screen)
+
         for obstacle in self.obstacles:
             obstacle.display_character(self.screen)
+        
 
     def calculate_score(self):
         # self.score += int((pygame.time.get_ticks() * self.game_speed**10) // 1000)
@@ -236,13 +241,13 @@ class Map(GameScreen):
 
     def checking_hp(self) -> bool:
         if (self.player1.hp == 2):
-            print('2')
+            # print('2')
             self.hp1 = pygame.image.load('textures/kosa.png').convert_alpha()
             self.hp1_rect = self.hp1.get_rect(center=(self.w - 80, self.h - 750))
             self.hp1 = pygame.transform.rotozoom(self.hp1, 0, 0.40)
             return False
         elif (self.player1.hp == 1):
-            print('1')
+            # print('1')
             self.hp2 = pygame.image.load('textures/kosa.png').convert_alpha()
             self.hp2_rect = self.hp2.get_rect(center=(self.w - 40, self.h - 750))
             self.hp2 = pygame.transform.rotozoom(self.hp2, 0, 0.40)
@@ -251,7 +256,7 @@ class Map(GameScreen):
             self.hp3 = pygame.image.load('textures/kosa.png').convert_alpha()
             self.hp3_rect = self.hp3.get_rect(center=(self.w, self.h - 750))
             self.hp3 = pygame.transform.rotozoom(self.hp3, 0, 0.40)
-            print('deat')
+            # print('deat')
             return True
 
     def show_life(self):
@@ -259,58 +264,4 @@ class Map(GameScreen):
         self.screen.blit(self.hp2, self.hp2_rect)
         self.screen.blit(self.hp3, self.hp3_rect)
 
-
-class Menu(GameScreen):
-    def __init__(self, w, h):
-        super().__init__(w, h)
-        # self.garage_button = pygame.image.load('textures/buttons/butgarage').convert_alpha()
-        # self.garage_button_rect = self.garage_button.get_rect(center = (300, 50))
-        # self.start_button = pygame.image.load('textures/buttons/butsound').convert_alpha()
-        # self.start_button_rect = self.start_button.get_rect(center = (300, 300))
-        # self.sound_button = pygame.image.load('textures/buttons/butstart').convert_alpha()
-        # self.sound_button_rect = self.sound_button.get_rect(center = (300, 550))
-
-        
-        self.text1 = self.font.render("press SPACE to start", True, 'black')
-        self.text_rect1 = self.text1.get_rect(center = (self.w//2, self.h//2))
-        self.text2 = self.font.render("press ESC to go to the garage", True, 'black')
-        self.text_rect2 = self.text2.get_rect(center = (self.w//2, self.h//2 - 200))
-
-        
-
-    def display_menu_bg(self):
-        self.screen.fill('lightblue')
-        self.screen.blit(self.text1, self.text_rect1)
-        self.screen.blit(self.text2, self.text_rect2)
-
-    def display_buttons(self):
-        self.screen.blit(self.garage_button, self.garage_button_rect)
-        self.sceen.blit(self.start_button, self.start_button_rect)
-        self.screen.blit(self.sound_button, self.sound_button_rect)
-
-    def click_button(self):
-        if self.garage_button_rect.collidepoint(pygame.mouse.get_pos()):
-            pass
-        # tu powinien zmienić się stan maszyny stanów na garage
-        if self.start_button_rect.collidepoint(pygame.mouse.get_pos()):
-            pass
-        # tu powinien zmienić się stan maszyny stanów na game (powinna się zacząć gra)
-        if self.sound_button_rect.collidepoint(pygame.mouse.get_pos()):
-            pass # tu powinien zmienić się stan muzyki na off ale nie ma jeszcze muzyki
-
-
-
-class Garage(GameScreen):
-    def __init__(self, w, h):
-        super().__init__(w, h)
-        self.text1 = self.font.render("press m to go back to menu", True, 'black')
-        self.text_rect1 = self.text1.get_rect(center = (self.w//2, self.h//2))
-        
-    def display_garage(self):
-        self.screen.fill('brown')
-        self.screen.blit(self.text1, self.text_rect1)
-
-
-
-
-
+
